Added ability to check last sumbitted emotion
